refactor(viewmodels): update namespaces for view model classes
Changed the namespace for AppModel, Drive, Sync, and User classes from kDrive_client.DataModel to kDrive_client.ViewModels to better reflect their roles in the application.
