+++ src/gui4/windows/kDrive client/kDrive client/ViewModels/AppViewModel.cs
@@ -29,7 +29,8 @@ using System.Threading.Tasks;
 using DynamicData;
 using DynamicData.Binding;
 
-namespace kDrive_client.DataModel
+namespace kDrive_client.ViewModels
+
 {
     internal class AppModel : ObservableObject
     {
