+++ src/gui4/windows/kDrive client/kDrive client/ViewModels/DriveViewModel.cs
@@ -28,7 +28,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
-namespace kDrive_client.DataModel
+namespace kDrive_client.ViewModels
 {
     internal class Drive : ObservableObject
     {
