+++ src/gui4/windows/kDrive client/kDrive client/ViewModels/UserViewModel.cs
@@ -30,7 +30,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace kDrive_client.DataModel
+namespace kDrive_client.ViewModels
 {
     internal class User : ObservableObject
     {
